pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "pillow>=12.0.0",
    "pygame>=2.6.1",
    "python-bidi>=0.6.7",
]
